Frontend: Major code refactor
cleaned up the code a lot (removed around 600 lines!)
I did that by turing the the card used to display a member or department
into it's own compnent 'leaderboard-card' and just reused that everywhere insted of
rewriting the same thing each time.

**comment**: this is a great first step into cleaning the code so that a
arabic translation in the future becomes much easier when the entire
site is just a limited set of components that can easily be edited.